Combine BlockBasedTrack and TrackBlocks to reduce redundancy
